+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostPollsRatingPlugin/PollRatingPlugin.kt
@@ -91,6 +91,11 @@ class PollRatingPlugin(
                 PostsTable
             )
 
+            enableReenableRatingCommand(
+                this@PollRatingPlugin,
+                PostsTable
+            )
+
             enableGetRatingsCommand(
                 executor.asReference(),
                 this@PollRatingPlugin