+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 ### 0.1.2
 
 * Now new polls will reply to first message of post
+* `reenableRating` command added